Don't crashed on FetchedData assertion: Closes #7146
Assertion is incorrect here because it's possible for the data
not to be returned for valid reasons -- usually a command response
failure (i.e. "NO").  Still want to log it for debugging purposes,
but don't crash when it happens.

Note that this patch also won't return the email if it doesn't
fulfill all the fields.