Force remote folder to open after timeout
A recent optimization for EmailStore was to only open a connection
to the remote folder if a command came in requiring it.  If an
ImapEngine.GenericFolder is opened for monitoring, however, no command
may come in requiring a remote connection, meaning no folder
normalization or unsolicited notifications arrive.

This patch "forces" the remote to open after a timeout, which is
cancelled if it's opened by a command.