No pauses while writing drafts: Closes #7426
The slight pause (sometimes up to 2s on my machine) was due to some
CPU-intensive work occurring during folder normalization.  Some of
the work can be done in a background thread (using the new
Nonblocking.Concurrent class), but some of the problem was simply
that the Deserializer's priority was too high, causing it to block
UI events when it was busy (such as requesting a lot of UIDs from
the server).