Load mail on ukr.net's Freemail server: Closes #7425
Two bugs are fixed in this patch:

(a) Freemail's XLIST returns the Inbox's name in Ukranian.  This
is a problem when doing a direct list (i.e. XLIST "" "Inbox") to
get folder properties.  This is fixed by using LIST whenever listing
Inbox.

(b) Freemail returns FETCH BODY[] header fields back in an
indeterminate order.  Previously the IMAP stack relied on a string
comparison to match up the response with the request.  New code
now properly (but painfully) parses the BODY[] response, obviating
the need for FetchBodyDataIdentifier.