More time for notification timing delay: Refs #6085
I noticed the double-archive problem again just now.  Looking at
the log, it took even longer than 250ms for Gmail to notify of the
EXPUNGEs after the append, so this patch increases the delay to a
full second.

It may be that this delay strategy is only pushing the timing hole
around, but until we can come up with a better solution, it'll have
to do.