Reduce local delays when new message arrives
In particular, bug #713493 reports this, although this patch doesn't
appear to solve the problem entirely.  However, I have spotted
situations in the past where the Append replay operation caused local
operations to hang.  This is because Append was being treated as a
local operation when, in fact, it's first call is back to the server
to fetch UIDs of the new messages.  Hence, it should be treated as
a remote operation so local operations can run without delay.