Unneccessary dconf updates on window move
Writing the window_{width,height,maximized} variables triggers
a dconf database update. This is an expensive operation as
dconf writes down a full copy of the database and renames
over the old one. dconf does not check for unchanged values.

configure_event() is called for every pixel of a window move,
and writes the window_* variables - those do not change during
a move. Check by running
"dconf watch /org/yorba/geary/" and moving the window or just
listen to the hard disk grinding.

This patch checks if the values have changed before writing
them out, which fixes the issue and provides a
disk-grinding-free window move experience.

Closes: bgo #724953