Generate and install localized help docs
This causes us to generate localized help docs whenever Geary builds,
and then the localized docs are installed with the rest of the help
pages.

It would be great to also have the CMakeLists read from the Makefile.am
file to determine what translations to run, but that hasn't happened
yet.

Closes: bgo #713831