Don't save draft when UIDPLUS is not present: Closes bgo#713983
Without UIDPLUS Geary currently cannot determine the UID of the drafts
it saves to the server, and so it cannot delete them when the draft is
saved again or the message is sent.  This patch simply checks if
UIDPLUS is supported, and if not, does not save the draft to the
server.

Although a string is included in this patch, it's a duplicate of an
existing string and so technically does not break string freeze.