Allow multiple inline composers
Replying to a message that already has a composer open will result in
that composer refocusing.  It should also adjust the Reply/Reply
All/Forward status, but that's still to come.

When switching away from a conversation with inline composers, the only
option is to cancel the operation or discard all compositions.  This
will be fixed by adding automatic saving of drafts in this case.