New scrolling system for inline composers
The editor in the composer no longer shows its own scrollbar.  Instead,
the conversation view allocates enough space to hold the composer
without any scrolling.  The composer then positions and scrolls itself
to create the illusion that the outer scroll bar controls it.

To track the size of the composer, we put all the text in a div and mark
that as contenteditable.  Then, when the div changes height, we update
the layout.  We use the user_changed_contents signal as a proxy for
this.

https://bugzilla.gnome.org/show_bug.cgi?id=730955