Only report Inbox deselected if Inbox is being removed: Bug #727744
Changing a specal folder type involves adding and removing the folder
entry in the folder sidebar.  Prior logic would look in the Inboxes
branch if the entry was not selected in the "normal" account branch
but not verify the folder in Inboxes was the one being removed.  This
triggered an unwarranted "nothing selected" signal that caused the
conversation list to go blank at startup when Geary was associating
special folder types with folders.

Note that you must have multiple accounts registered w/ Geary to
trigger this bug.