Support multiple Message-IDs with In-Reply-To
Prior RFC822 code treated In-Reply-To as only capable of listing a
single Message-ID parent.  The spec does allow for multiple parent
messages, so this change reflects that.

This potentially can also fix bugs where messages were not included
in conversations, but this would be an edge case, as that would
require the MUA not listing the In-Reply-To Message-IDs in the
References list, which Geary does properly parse.