Present Welcome dialog if Geary re-run while visible: Bug #720360
If the Welcome dialog (used to create the initial account) is visible
and Geary was re-run, an empty main window would appear.  This is due
to the Welcome dialog hiding the main window until the first account
is created.  Now, if Geary is re-run the Welcome dialog is presented
to the user.

This may possibly solve bug #737811.